+++ README.md
@@ -23,13 +23,11 @@
 # Step 3.
 ## 빼빼로 그리기
 
-빼빼로 몸통 부분을 만들어주는 함수 : getLineOfPepero()
+빼빼로의 정보를 출력해주는 함수 : printInformOfPepero()
 빼빼로 몸통을 그려주는 함수 : drawBodyOfPepero()
 빼빼로 스틱 부분을 그려주는 함수 : drawStickOfPepero()
 
-### 의문점:
-    - "Step 2에 작성했던 Command Line Tool 프로젝트 내부에 새로운 파일을 생성합니다." 부분은 제가 한 것처럼 하면 되는 걸까요?
-        + 함수를 PeperoFile에 작성하고 main에서 함수를 호출했습니다.
-
-    - 스틱 부분을 출력할 때 한줄 띄워놓았는데 제가 제대로 이해한 것 일까요?
-        + 조금 아쉬운 부분은 몸통 넓이에 따라 스틱이 자유자재로 위치를 달리해서 출력해주었으면 하네요...
+### 수정:
+    - 몸통 부분을 만들어주는 함수가 필요 없어보여 없애고 그 부분을 몸통 그려주는 함수에 넣었습니다!
+    - 빼빼로의 정보를 출력해주는 함수 printInformOfPepero()를 만들었습니다.
+    - 매개변수 등 때문에 변수 이름이 계속 만들어지는 것 같아 해결 방법을 찾다가 (_ 매개변수 이름: 타입)으로 해결해보았습니다.
